Add middleware authorization

+++ routes.php
@@ -1,22 +1,25 @@
 <?php
 
+use JWord\Framework\Middleware\Authorize;
+
 // GET listing routes
 $router->get('/workopia/', 'HomeController@index');
 $router->get('/workopia/listings', 'ListingController@index');
-$router->get('/workopia/listings/create', 'ListingController@create');
-$router->get('/workopia/listings/edit/{id}', 'ListingController@edit');
+$router->get('/workopia/listings/create', 'ListingController@create', ['auth']);
+$router->get('/workopia/listings/edit/{id}', 'ListingController@edit', ['auth']);
 $router->get('/workopia/listings/{id}', 'ListingController@show');
 
+
 // POST listing routes
-$router->post('/workopia/listings', 'ListingController@store');
-$router->put('/workopia/listings/{id}', 'ListingController@update');
-$router->delete('/workopia/listings/{id}', 'ListingController@destroy');
+$router->post('/workopia/listings', 'ListingController@store', ['auth']);
+$router->put('/workopia/listings/{id}', 'ListingController@update', ['auth']);
+$router->delete('/workopia/listings/{id}', 'ListingController@destroy', ['auth']);
 
 // GET user routes
-$router->get('/workopia/auth/login', 'UserController@login');
-$router->get('/workopia/auth/register', 'UserController@create');
+$router->get('/workopia/auth/login', 'UserController@login', ['guest']);
+$router->get('/workopia/auth/register', 'UserController@create', ['guest']);
 
 // POST user routes
-$router->post('/workopia/auth/register', 'UserController@store');
-$router->post('/workopia/auth/logout', 'UserController@logout');
-$router->post('/workopia/auth/login', 'UserController@authenticate');
+$router->post('/workopia/auth/register', 'UserController@store', ['guest']);
+$router->post('/workopia/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/workopia/auth/login', 'UserController@authenticate', ['guest']);
